feat(dui): wip moving the one click send in DUI
