+++ DesktopUI2/DesktopUI2/Models/ConfigManager.cs
@@ -33,6 +33,7 @@ namespace DesktopUI2.Models
   public class Config
   {
     public bool DarkTheme { set; get; }
+    public bool OneClickMode { set; get; } = true;
     public bool ShowImportExportAlert { set; get; } = true;
   }
