+++ DesktopUI2/DesktopUI2/Views/Controls/SavedStreams.xaml.cs
@@ -2,11 +2,11 @@ using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
 
-namespace DesktopUI2.Views.Pages.StreamEditControls
+namespace DesktopUI2.Views.Controls
 {
-  public partial class Send : UserControl
+  public partial class SavedStreams : UserControl
   {
-    public Send()
+    public SavedStreams()
     {
       InitializeComponent();
     }
