+++ DesktopUI2/DesktopUI2/Views/Controls/StreamDetails.xaml
@@ -1,5 +1,5 @@
 <UserControl
-  x:Class="DesktopUI2.Views.Pages.SharedControls.StreamDetails"
+  x:Class="DesktopUI2.Views.Controls.StreamDetails"
   xmlns="https://github.com/avaloniaui"
   xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
   xmlns:conv="clr-namespace:DesktopUI2.Views.Converters"
