+++ DesktopUI2/DesktopUI2/Views/Controls/StreamEditControls/Activity.xaml.cs
@@ -1,11 +1,12 @@
+using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
 
-namespace DesktopUI2.Views.Pages.StreamEditControls
+namespace DesktopUI2.Views.Controls.StreamEditControls
 {
-  public partial class Comments : UserControl
+  public partial class Activity : UserControl
   {
-    public Comments()
+    public Activity()
     {
       InitializeComponent();
     }
