+++ DesktopUI2/DesktopUI2/Views/Controls/StreamEditControls/Comments.xaml.cs
@@ -1,12 +1,11 @@
-using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
 
-namespace DesktopUI2.Views.Pages.StreamEditControls
+namespace DesktopUI2.Views.Controls.StreamEditControls
 {
-  public partial class Report : UserControl
+  public partial class Comments : UserControl
   {
-    public Report()
+    public Comments()
     {
       InitializeComponent();
     }
