+++ DesktopUI2/DesktopUI2/Views/Controls/StreamEditControls/Receive.xaml.cs
@@ -2,11 +2,11 @@ using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
 
-namespace DesktopUI2.Views.Pages.StreamEditControls
+namespace DesktopUI2.Views.Controls.StreamEditControls
 {
-  public partial class Activity : UserControl
+  public partial class Receive : UserControl
   {
-    public Activity()
+    public Receive()
     {
       InitializeComponent();
     }
