+++ DesktopUI2/DesktopUI2/Views/Controls/StreamEditControls/Report.xaml
@@ -1,5 +1,5 @@
 <UserControl
-  x:Class="DesktopUI2.Views.Pages.StreamEditControls.Report"
+  x:Class="DesktopUI2.Views.Controls.StreamEditControls.Report"
   xmlns="https://github.com/avaloniaui"
   xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
   xmlns:assists="clr-namespace:Material.Styles.Assists;assembly=Material.Styles"
