+++ DesktopUI2/DesktopUI2/Views/Pages/CollaboratorsView.xaml.cs
@@ -1,9 +1,9 @@
 using Avalonia.Markup.Xaml;
 using Avalonia.ReactiveUI;
-using DesktopUI2.ViewModels.Share;
+using DesktopUI2.ViewModels;
 using ReactiveUI;
 
-namespace DesktopUI2.Views.Pages.ShareControls
+namespace DesktopUI2.Views.Pages
 {
   public partial class CollaboratorsView : ReactiveUserControl<CollaboratorsViewModel>
   {
