--- DesktopUI2/DesktopUI2/Views/Pages/ShareControls/Sending.xaml
@@ -1,119 +1,0 @@
-<UserControl
-  x:Class="DesktopUI2.Views.Pages.ShareControls.Sending"
-  xmlns="https://github.com/avaloniaui"
-  xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-  xmlns:assists="clr-namespace:Material.Styles.Assists;assembly=Material.Styles"
-  xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-  xmlns:icons="clr-namespace:Material.Icons.Avalonia;assembly=Material.Icons.Avalonia"
-  xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-  assists:ShadowAssist.ShadowDepth="10"
-  d:DesignHeight="250"
-  d:DesignWidth="600"
-  Background="{DynamicResource PrimaryHueMidBrush}"
-  mc:Ignorable="d">
-  <Grid>
-    <Grid IsVisible="{Binding SendComplete, Converter={x:Static BoolConverters.Not}, FallbackValue=False}" RowDefinitions="auto, auto, auto, auto">
-      <TextBlock
-        Margin="15,15,15,0"
-        Classes="Subtitle1"
-        Foreground="White"
-        Text="{Binding Progress.ProgressTitle, FallbackValue='Operation X'}"
-        TextTrimming="CharacterEllipsis" />
-      <ProgressBar
-        Grid.Row="1"
-        Margin="15,15,15,0"
-        IsIndeterminate="{Binding Progress.IsIndeterminate, FallbackValue=True}"
-        IsVisible="{Binding Progress.IsProgressing}"
-        Maximum="{Binding Max}"
-        Value="{Binding Value}" />
-      <TextBlock
-        Grid.Row="2"
-        Margin="15,15,15,0"
-        Foreground="White"
-        Text="{Binding Progress.ProgressSummary, FallbackValue='Operation in progress...'}"
-        TextTrimming="CharacterEllipsis" />
-      <StackPanel
-        Grid.Row="3"
-        Margin="15"
-        HorizontalAlignment="Right"
-        Orientation="Horizontal">
-        <Button
-          Margin="0,0,10,0"
-          Classes="Outline"
-          Command="{Binding Progress.CancelCommand}"
-          Content="Cancel"
-          Foreground="White"
-          IsVisible="{Binding Progress.IsProgressing}" />
-        <!--<Button
-        Margin="0,0,10,0"
-        Classes="Outline"
-        Command="{Binding Progress.OpenReportCommand}"
-        Content="Open Report"
-        Foreground="White"
-        IsVisible="{Binding Progress.IsProgressing, Converter={x:Static BoolConverters.Not}}" />-->
-        <!--<Button
-        Margin="0,0,10,0"
-        Background="White"
-        Click="Close_Click"
-        Content="Close"
-        Foreground="{DynamicResource PrimaryHueMidBrush}"
-        IsEnabled="{Binding IsProgressing, Converter={x:Static BoolConverters.Not}}" />-->
-      </StackPanel>
-    </Grid>
-
-    <Grid
-      HorizontalAlignment="Center"
-      IsVisible="{Binding SendComplete, FallbackValue=True}"
-      RowDefinitions="auto, auto, auto, auto">
-      <TextBlock
-        Margin="30"
-        HorizontalAlignment="Center"
-        Classes="Subtitle1"
-        Foreground="White"
-        Text="Done 🎉! Your model has been shared successfully"
-        TextTrimming="CharacterEllipsis" />
-      <Button
-        Grid.Row="1"
-        Margin="0"
-        HorizontalAlignment="Center"
-        CornerRadius="25"
-        Background="White"
-        Command="{Binding ViewOnlineCommand}"
-        Content="Cancel"
-        Foreground="{DynamicResource PrimaryHueMidBrush}">
-        <Grid Margin="0" ColumnDefinitions="auto,*">
-
-          <icons:MaterialIcon VerticalAlignment="Center" Kind="ExternalLink" />
-          <TextBlock
-            Grid.Column="1"
-            Margin="8,0"
-            Padding="0"
-            VerticalAlignment="Center"
-            Text="View online"
-            TextTrimming="CharacterEllipsis" />
-        </Grid>
-      </Button>
-      <Button
-        Grid.Row="2"
-        Margin="0,20,0,40"
-        HorizontalAlignment="Center"
-        CornerRadius="25"
-        Classes="Flat"
-        Command="{Binding CopyCommand}"
-        Content="Cancel"
-        Foreground="White">
-        <Grid Margin="0" ColumnDefinitions="auto,*">
-
-          <icons:MaterialIcon VerticalAlignment="Center" Kind="LinkVariant" />
-          <TextBlock
-            Grid.Column="1"
-            Margin="8,0"
-            Padding="0"
-            VerticalAlignment="Center"
-            Text="Copy URL to clipboard"
-            TextTrimming="CharacterEllipsis" />
-        </Grid>
-      </Button>
-    </Grid>
-  </Grid>
-</UserControl>
