--- DesktopUI2/DesktopUI2/Views/Share.xaml
@@ -1,29 +1,0 @@
-﻿<Window
-  x:Class="DesktopUI2.Views.Share"
-  xmlns="https://github.com/avaloniaui"
-  xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-  xmlns:assists="clr-namespace:Material.Styles.Assists;assembly=Material.Styles"
-  xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-  xmlns:icons="clr-namespace:Material.Icons.Avalonia;assembly=Material.Icons.Avalonia"
-  xmlns:m="clr-namespace:Material.Styles;assembly=Material.Styles"
-  xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-  xmlns:p="clr-namespace:DesktopUI2.Views.Pages"
-  xmlns:reactiveUi="http://reactiveui.net"
-  xmlns:vm="using:DesktopUI2.ViewModels.DesignViewModels"
-  Title="{Binding TitleFull}"
-  Width="500"
-  Icon="/Assets/icon.ico"
-  SizeToContent="Height"
-  mc:Ignorable="d">
-  <Design.DataContext>
-    <vm:DesignShareViewModel />
-  </Design.DataContext>
-
-
-  <reactiveUi:RoutedViewHost Grid.Row="1" Router="{Binding Router}">
-    <reactiveUi:RoutedViewHost.PageTransition>
-      <CrossFade Duration="0.15" />
-    </reactiveUi:RoutedViewHost.PageTransition>
-  </reactiveUi:RoutedViewHost>
-
-</Window>
