+++ DesktopUI2/DesktopUI2/Views/Windows/Dialogs/ChangeRoleDialog.xaml
@@ -1,5 +1,5 @@
 <UserControl
-  x:Class="DesktopUI2.Views.Pages.ShareControls.ChangeRoleDialog"
+  x:Class="DesktopUI2.Views.Windows.Dialogs.ChangeRoleDialog"
   xmlns="https://github.com/avaloniaui"
   xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
   xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
