+++ DesktopUI2/DesktopUI2/Views/Windows/Dialogs/ChangeRoleDialog.xaml.cs
@@ -1,9 +1,8 @@
 using Avalonia.Controls;
 using Avalonia.Interactivity;
 using Avalonia.Markup.Xaml;
-using DesktopUI2.Views.Windows.Dialogs;
 
-namespace DesktopUI2.Views.Pages.ShareControls
+namespace DesktopUI2.Views.Windows.Dialogs
 {
   public partial class ChangeRoleDialog : DialogUserControl
   {
@@ -30,12 +29,12 @@ namespace DesktopUI2.Views.Pages.ShareControls
     {
       //too lazy to create a view model for this or properly style the Dialogs
       Role = (RoleField.SelectedItem as ComboBoxItem).Content as string;
-      this.Close(Role);
+      Close(Role);
     }
 
     public void Close_Click(object sender, RoutedEventArgs e)
     {
-      this.Close(null);
+      Close(null);
     }
   }
 }
